requirements.txt
pandas>=2.0.0
scikit-learn>=1.0.0
matplotlib>=3.0.0
google-generativeai>=0.1.0